updates index script
